+++ CHANGELOG.md
@@ -1,8 +1,13 @@
+### Upcoming
+
+ - Updated the Export Data page to select a single experiment at a time. This reduces the load on the backend which was causing the system to crash for large exports.
+ - Also improved the performance of exports to export larger datasets. We also added safety checks during the export process to monitor the system and abort if it gets too close to some memory and size thresholds.
+
 ### 26.4.4
 
 #### Enhancements
-
  - New multi-pioreactor flow in the UI to run simultaneous protocols for stirring calibrations.
+
 
 #### Breaking changes